% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abie.R
\docType{package}
\name{abie}
\alias{abie}
\alias{abie-package}
\title{abie: a package for assay-based incidence estimation}
\description{
The abie package uses methods described by Kassanjee, et al. "A new general biomarker-based incidence estimator," \emph{Epidemiology} (2012), to implement functions to calculate incidence and tests of incidence difference between two populations, as well as power and sample size constraints for differenct study scenarios.
abie also provides functions for calculation of mean duration of recent infection and false recency rates from assays for recent infection.
}
\section{abie Functions}{

abie has functions \emph{frr_binomial} to estimate false recency rate; \emph{mdri_ml_binomial} to estimate mean duration or recent infection; \emph{recencyI} to calculate estimates and confidence intervals for incidence and incidence difference, as well as other summary and inferential statistics related to the survey; \emph{SSPower} to calculate sample size needed for a given power in a test of incidence difference, or vice versa; and \emph{SSCprecision}, which gives sample size needed for a given precision in the incidence estimate.
}

