% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recencyI_prev.R
\docType{data}
\name{probs}
\alias{probs}
\title{Incidence and incidence difference statistics from trinomial prevalences of HIV and recency}
\format{An object of class \code{data.frame} with 3 rows and 4 columns.}
\usage{
probs
}
\arguments{
\item{BS_Count}{Specifies number of bootstrap samples for bootstrapped confidence intervals of incidence.}

\item{BS_Vars}{True/False variable indicating whether variance of point estimates is to be calculated by Empirical Bootstrapping (TRUE) or Delta Method (FALSE), the default setting.}

\item{BMest}{Biomarker estimation by one the 3 options "same.test"(=default), "FRR.indep", "MDRI.FRR.idep" (string).}

\item{PrevH}{Prevelance of HIV in survey.}

\item{RSE_PrevH}{Relative Standard Error (RSE) of estimate for population prevalence of HIV}

\item{PrevR}{Proportion of persons found to be 'recent' by biomarker assay among total persons found positive for HIV.}

\item{RSE_PrevR}{Relative Standard Error (RSE) of estimate for population proportion of those testing positive for HIV who have been infected recently.}

\item{MDRI}{Mean Duration of Recent Infection as measured by a biomarker assay to test recency among positive cases.}

\item{RSE_MDRI}{Relative Standard Error (RSE) of parameter MDRI as measured by the biomarker assay to test recency in the given prevalence survey.}

\item{FRR}{False Recency Rate as measured by the biomarker assay to test recency in the given prevalence survey.}

\item{RSE_FRR}{Relative Standard Error (RSE) of parameter FRR as measured by the biomarker assay to test recency in the given prevalence survey.}

\item{BigT}{Cut point in days of recency used in biomarker assay to test recency in a given prevalence survey.}

\item{Covar_HR}{Covariance of probability of being postive and being categorized recent from survey (or as a vector for multiple surveys)}
}
\value{
Incidence

RSE Incidence

etc.
}
\description{
Incidence and incidence difference statistics from trinomial prevalences of HIV and recency
}
\details{
general details
}
\examples{
example
}
\keyword{datasets}

