% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recencyI_prev.R
\name{recencyI}
\alias{recencyI}
\title{Incidence and incidence difference from Trinomial Counts / Prevalences of HIV and recency}
\usage{
recencyI(BS_Count = 10000, BSDM_spread = 1000, BS_Vars = NULL,
  BMest = "sameTest", PrevH, RSE_PrevH, PrevR, RSE_PrevR, MDRI, RSE_MDRI, FRR,
  RSE_FRR, BigT = 730, Covar_HR = 0)
}
\arguments{
\item{BS_Count}{Add details on each prarameter}
}
\value{
Incidence

RSE Incidence

etc.
}
\description{
Incidence and incidence difference from Trinomial Counts / Prevalences of HIV and recency
}
\details{
general details
}
\examples{
example
}

