% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSCPrecision_VaryPrelim.R
\name{SSCprecision}
\alias{SSCprecision}
\title{Sample Size or Precision Calculation}
\usage{
SSCprecision(I, RSE_I, PrevH, CR, MDRI, RSE_MDRI, FRR, RSE_FRR, BigT = 730,
  DE_H = 1, DE_R = 1, n = "out", step = 5)
}
\arguments{
\item{I}{Expected Incidence}

\item{RSE_I}{Relative Standard Error of Incidence Estimate}

\item{PrevH}{Prevalence of HIV}

\item{CR}{Coverage rate: probability (0-1) of being tested for recency when positive for HIV}

\item{MDRI}{mean duration of recent infection [days] (vector/integer)}

\item{RSE_MDRI}{Relative standard error of MDRI   [days] (vector/integer)}

\item{FRR}{False recent rate (vector/integer)}

\item{RSE_FRR}{Relative standard error of FRR (vector/integer)}

\item{BigT}{post-infection time cut-off true vs false recent [days] default 730 days (integer)}

\item{DE_H}{Design effect of HIV-prevalence test (vector/integer)}

\item{DE_R}{Design effect of recency test (vector/integer)}

\item{n}{Sample Size, either a given hypothetical value or to be determined by function (default)}

\item{step}{number of steps between minimum I and maximum I in the calculation of a range of output}
}
\value{
Either sample size necessary for a given precision under a given set of testing characteristics and a hypothetical prevalence/incidence scenario, or precision under a particular sample size scenario, with a given hypothetical prevalence/incidence scenario.
}
\description{
Sample Size or Precision Calculation
}
\details{
Summarizes performance of a recent infection test (into a standard error of the incidence estimate), given estimated test properties (RSE of incidence) and the prevalence/incidence in a hypothetical context; or gives sample size necessary for a given level of estimator precision.
Returns: proportion of sample categorized as HIV positive and recently infected; proportion of sample categorized as HIV positive and non-recently infected; the relative standard error of the incidence estimator at infinite sample size, which is the component of variability explained soley by the assay characteristics; the relative standard error of the estimate of prevalence; the relative standard error of the estimate of proportion of HIV positive that are recent.
}
\examples{
SSCprecision(I = 0.015, RSE_I = 0.25, PrevH = 0.2, CR = 1,
MDRI = 200, RSE_MDRI = 0.05, FRR = 0.01, RSE_FRR = 0.2,
BigT = 730, DE_H = 1.1, DE_R = 1, n = "out")

SSCprecision(I = c(0.015,0.02), RSE_I = 0.25, PrevH = c(0.10,0.20),
CR = 1, MDRI = 200, RSE_MDRI = 0.05, FRR = 0.01, RSE_FRR = 0.2,
BigT = 700, DE_H = 1, DE_R = 1, n = "out", step = 5)

SSCprecision(I = 0.017, RSE_I = "out", PrevH = c(0.10,0.20),
CR = 1, MDRI = 211, RSE_MDRI = 0.05, FRR = 0.009, RSE_FRR = 0.2,
BigT = 720, n = 5000, step = 5)
}

