% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSPower.R
\name{SSPower}
\alias{SSPower}
\title{Power and Sample Size calculation for assay-based incidence estimation}
\usage{
SSPower(I1, I2, PrevH1, PrevH2, n1, n2, alpha = 0.05, Power = 0.8,
  SS = "out", CR = 1, DE_H = 1, DE_R = 1, BMest = "same.test", MDRI,
  RSE_MDRI, FRR, RSE_FRR, BigT = 730)
}
\arguments{
\item{I1}{Predicted incidence of HIV in survey 1}

\item{I2}{Predicted incidence of HIV in survey 2}

\item{PrevH1}{Predicted prevalence of HIV in survey 1}

\item{PrevH2}{Predicted prevalence of HIV in survey 2}

\item{n1}{Sample size for survey 1}

\item{n2}{Sample size for survey 2}

\item{alpha}{Significance level for test (default alpha=0.05)}

\item{Power}{Desired power used to calculate a sample size for the surveys. Default is 0.80, meaning the function outputs the necessary sample size to achieve stated power for a test of differences in incidence.}

\item{SS}{Sample size. Default is NULL, meaning the function takes a power argument and outputs sample size needed to achieve power level for test of differences for incidence. If parameter is set to a number, then function will return the power achieved by the test of differences under that sample size.}

\item{DE_H}{Design effect of HIV-prevalence test (vector/integer)}

\item{DE_R}{Design effect of recency test (vector/integer)}

\item{BMest}{Biomarker estimation by one the 3 options "same.test"(=default), "FRR.indep", "MDRI.FRR.indep" (string)}

\item{MDRI}{mean duration of recent infection [days] (vector/integer)}

\item{RSE_MDRI}{Relative standard error of MDRI [days] (vector/integer)}

\item{FRR}{False recent rate (vector/integer)}

\item{RSE_FRR}{Relative standard error of FRR (vector/integer)}

\item{BigT}{post-infection time cut-off true vs false recent [days] default 730 days (integer)}
}
\value{
Sample size necessary for a given power level for testing the hypothesis that the incidence rates are identical between populations; alternatively, the power of said test under a particular sample size scenario. Function also returns implied statistics from input values, confidence limits, and population counts.
}
\description{
Power and Sample Size calculation for assay-based incidence estimation
}
\details{
...
}
\examples{
SSPower(I1=0.05, I2=0.03, PrevH1=0.20, PrevH2=0.20,
n1=5000, n2=5000, alpha=0.05, Power="out", SS=NULL,
DE_H=1, DE_R=1, BMest="same.test", MDRI=200,
RSE_MDRI=0.05, FRR=0.01, RSE_FRR=0.20, BigT=730)


SSPower(I1=0.05, I2=0.03, PrevH1=0.20, PrevH2=0.20,
alpha=0.05, Power=0.80, SS="out", DE_H=1, DE_R=1,
BMest="FRR.indep", MDRI=200, RSE_MDRI=0.05, FRR=0.01,
RSE_FRR=0.20, BigT=730)
}

