% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recencyI_prev.R
\name{incBYcounts}
\alias{incBYcounts}
\title{Incidence and incidence difference statistics from trinomial counts of HIV and recency}
\usage{
incBYcounts(N, N_H, N_testR, N_R, DE_H = 1, DE_R = 1, BS_Count = 10000,
  Boot = FALSE, BMest = "same.test", MDRI, RSE_MDRI, FRR, RSE_FRR,
  BigT = 730, Covar_HR = 0)
}
\arguments{
\item{N}{Total sample size of survey}

\item{N_H}{Count of persons testing positive for HIV}

\item{N_testR}{Count of persons tested for recency among those found to be HIV positive}

\item{N_R}{Count of persons found to be recently infected}

\item{DE_H}{Design effect of HIV-prevalence test (vector/integer)}

\item{DE_R}{Design effect of recency test (vector/integer)}

\item{BS_Count}{Specifies number of bootstrap samples for bootstrapped confidence intervals of incidence.}

\item{Boot}{True/False variable indicating whether variance of point estimates is to be calculated by Empirical Bootstrapping (TRUE) or Delta Method (FALSE), the default setting}

\item{BMest}{Biomarker estimation by one the 3 options "same.test"(=default), "FRR.indep", "MDRI.FRR.idep" (string)}

\item{MDRI}{mean duration of recent infection [days] (vector/integer)}

\item{RSE_MDRI}{Relative standard error of MDRI [days] (vector/integer)}

\item{FRR}{False recent rate (vector/integer)}

\item{RSE_FRR}{Relative standard error of FRR (vector/integer)}

\item{BigT}{Cut point in days of recency used in biomarker assay to test recency in a given prevalence survey}

\item{Covar_HR}{Covariance of probability of being postive and being categorized recent from survey (or as a vector for multiple surveys)}
}
\value{
Incidence estimate, confidence interval, relative standard error. If multiple surveys are entered, function returns identical results, as well as estimates of indcidence differences, confidence intervals of differences, differences in relative standard error, and p-values testing the hypothesis that the indcidence measures are different
}
\description{
Incidence and incidence difference statistics from trinomial counts of HIV and recency
}
\details{
Here is where we give a description in words of what the function does
}
\examples{
incBYcounts(5000 ,N_H = 1000, N_testR = 1000, N_R = 70, BS_Vars=TRUE, BMest="MDRI.FRR.indep",
MDRI = 200, RSE_MDRI = 0.05, FRR = 0.01, RSE_FRR = 0.2, BigT = 730)
}

