% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recencyI_prev.R
\name{recencyI}
\alias{recencyI}
\title{Incidence and incidence difference statistics from trinomial prevalences of HIV and recency}
\usage{
recencyI(PrevH, RSE_PrevH, PrevR, RSE_PrevR, Boot = FALSE, BS_Count = 10000,
  alpha = 0.05, BMest = "same.test", MDRI, RSE_MDRI, FRR, RSE_FRR,
  BigT = 730, Covar_HR = 0)
}
\arguments{
\item{PrevH}{Prevalence of HIV}

\item{RSE_PrevH}{Relative Standard Error (RSE) of estimate for population prevalence of HIV}

\item{PrevR}{Proportion of persons found to be 'recent' by biomarker assay among total persons found positive for HIV}

\item{RSE_PrevR}{Relative Standard Error (RSE) of estimate for population proportion of those testing positive for HIV who have been infected recently}

\item{Boot}{True/False variable indicating whether variance of point estimates is to be calculated by Empirical Bootstrapping (TRUE) or Delta Method (FALSE), the default setting}

\item{BS_Count}{Specifies number of bootstrap samples for bootstrapped confidence intervals of incidence}

\item{BMest}{Biomarker estimation by one the 3 options "same.test"(=default), "FRR.indep", "MDRI.FRR.idep" (string)}

\item{MDRI}{mean duration of recent infection [days] (vector/integer)}

\item{RSE_MDRI}{Relative standard error of MDRI [days] (vector/integer)}

\item{FRR}{False recent rate (vector/integer)}

\item{RSE_FRR}{Relative standard error of FRR (vector/integer)}

\item{BigT}{post-infection time cut-off true vs false recent [days] default 730 days (integer)}

\item{Covar_HR}{Covariance of probability of being postive and being categorized recent from survey (or as a vector for multiple surveys)}
}
\value{
Incidence estimate, confidence interval, relative standard error. If multiple surveys are entered, function returns identical results, as well as estimates of indcidence differences, confidence intervals of differences, differences in relative standard error, and p-values testing the hypothesis that the indcidence measures are different
}
\description{
Incidence and incidence difference statistics from trinomial prevalences of HIV and recency
}
\details{
general details
}
\examples{
recencyI(BS_Count=10000, Boot=FALSE, BMest="same.test", PrevH=0.20,
RSE_PrevH=0.028, PrevR=0.10, RSE_PrevR=0.094, MDRI=200, RSE_MDRI=0.05,
FRR=0.01, RSE_FRR=0.2, BigT=730)
}

